refactor(ui): dedupe kanban test fixtures
