+++ packages/ui/src/kanban-board/KanbanBoard.keyboard.test.tsx
@@ -1,14 +1,8 @@
 // @vitest-environment jsdom
 
-import type {
-  AppSettings,
-  GitHubIssueCacheRecord,
-  IssuePipelineStatus,
-  Project,
-} from '@shipcode/shared';
+import type { AppSettings, IssuePipelineStatus } from '@shipcode/shared';
 import { DEFAULT_SETTINGS } from '@shipcode/shared';
-import { act, type ReactElement, type ReactNode } from 'react';
-import { createRoot } from 'react-dom/client';
+import { act, type ReactNode } from 'react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 vi.mock('@dnd-kit/core', async () => {
@@ -35,118 +29,7 @@ vi.mock('@dnd-kit/core', async () => {
 });
 
 import { KanbanBoard } from '../KanbanBoard';
-
-function makeIssue(overrides: Partial<GitHubIssueCacheRecord> = {}): GitHubIssueCacheRecord {
-  return {
-    id: 'issue-1',
-    projectId: 'project-1',
-    issueNumber: 10,
-    title: 'Todo top',
-    body: null,
-    labels: [],
-    assignee: null,
-    state: 'open',
-    pipelineStatus: 'todo',
-    threadId: null,
-    claimedAt: null,
-    claimedBy: null,
-    lastPhaseUpdate: null,
-    lastStatusLabel: null,
-    plannerModelOverride: null,
-    reviewerModelOverride: null,
-    executorModelOverride: null,
-    verifierModelOverride: null,
-    plannerModelIdOverride: null,
-    reviewerModelIdOverride: null,
-    executorModelIdOverride: null,
-    verifierModelIdOverride: null,
-    plannerReasoningEffortOverride: null,
-    reviewerReasoningEffortOverride: null,
-    executorReasoningEffortOverride: null,
-    verifierReasoningEffortOverride: null,
-    revisionCountOverride: null,
-    requireApprovalOverride: null,
-    linkedPrNumber: null,
-    linkedPrUrl: null,
-    linkedPrIsDraft: false,
-    ciBlocked: false,
-    failingChecks: [],
-    unresolvedReviewComments: [],
-    unresolvedReviewCommentCount: 0,
-    prLastSyncAt: null,
-    fetchedAt: '2026-04-28T00:00:00.000Z',
-    priorityRank: null,
-    priorityRaw: null,
-    priorityFetchedAt: null,
-    isQuickMode: false,
-    ...overrides,
-  };
-}
-
-function makeProject(overrides: Partial<Project> = {}): Project {
-  return {
-    id: 'project-1',
-    name: 'ShipCode',
-    path: '/tmp/shipcode',
-    gitRemote: 'git@github.com:shipshitdev/shipcode.git',
-    githubRepoId: null,
-    githubRepoFullName: null,
-    starterIssueNumber: null,
-    starterIssueCreatedAt: null,
-    githubProjectUrl: null,
-    plannerModelOverride: null,
-    reviewerModelOverride: null,
-    executorModelOverride: null,
-    verifierModelOverride: null,
-    plannerModelIdOverride: null,
-    reviewerModelIdOverride: null,
-    executorModelIdOverride: null,
-    verifierModelIdOverride: null,
-    plannerReasoningEffortOverride: null,
-    reviewerReasoningEffortOverride: null,
-    executorReasoningEffortOverride: null,
-    verifierReasoningEffortOverride: null,
-    revisionCountOverride: null,
-    requireApprovalOverride: null,
-    discordRouting: 'inherit',
-    discordWebhookUrlOverride: null,
-    telegramRouting: 'inherit',
-    telegramChatIdOverride: null,
-    defaultBranch: 'main',
-    pinned: false,
-    archived: false,
-    hidden: false,
-    notifyGithubUser: null,
-    createdAt: '2026-04-28T00:00:00.000Z',
-    updatedAt: '2026-04-28T00:00:00.000Z',
-    ...overrides,
-  };
-}
-
-function renderIntoDom(element: ReactElement) {
-  const container = document.createElement('div');
-  document.body.appendChild(container);
-  const root = createRoot(container);
-
-  act(() => {
-    root.render(element);
-  });
-
-  return {
-    container,
-    rerender: (nextElement: ReactElement) => {
-      act(() => {
-        root.render(nextElement);
-      });
-    },
-    cleanup: () => {
-      act(() => {
-        root.unmount();
-      });
-      container.remove();
-    },
-  };
-}
+import { makeIssue, makeProject, renderIntoDom } from './test-helpers';
 
 function focusedCard(container: HTMLElement): HTMLElement {
   const card = container.querySelector('[data-keyboard-focused="true"]');
